Refactor: API ahora usa DTOs en lugar de entidades en Producto y Categoria

+++ gestion-productos/src/main/java/com/proyecto/gestion/controlador/ProductoControlador.java
@@ -4,13 +4,19 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.proyecto.gestion.dto.ProductoDTO;
 import com.proyecto.gestion.entidad.Producto;
+import com.proyecto.gestion.repositorio.ProductoRepositorio;
 import com.proyecto.gestion.servicio.ProductoServicio;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -40,7 +46,7 @@ public class ProductoControlador {
 	//Listar Id
 	@GetMapping("producto/{id}")
 	@ResponseStatus(HttpStatus.OK)
-	public List<Producto> todosProductosId(@PathVariable Integer id){
+	public ProductoDTO todosProductosId(@PathVariable Integer id){
 		
 		return (List<Producto>) productoServicio.buscarProductoPorId(id);
 		
@@ -50,7 +56,7 @@ public class ProductoControlador {
 	//Agregar
 	@PostMapping("producto")
 	@ResponseStatus(HttpStatus.CREATED)
-	public Producto nuevoProducto(@RequestBody Producto producto) {
+	public Producto nuevoProducto(@RequestBody ProductoDTO producto) {
 		
 		return productoServicio.agregarProducto(producto);
 		
@@ -62,9 +68,9 @@ public class ProductoControlador {
 	//Actualizar
 	@PutMapping("producto/{id}")
 	@ResponseStatus(HttpStatus.CREATED)
-	public Producto actualizarProducto(@PathVariable Integer id, @RequestBody Producto productoActualizado) {
+	public Producto actualizarProducto(@PathVariable Integer id, @RequestBody ProductoDTO productoActualizado) {
 		
-		Producto productoExistente = productoServicio.buscarProductoPorId(id);
+		ProductoDTO productoExistente = productoServicio.buscarProductoPorId(id);
 		
 		if(productoExistente != null) {
 			
@@ -72,7 +78,7 @@ public class ProductoControlador {
 			productoExistente.setNombre_prod(productoActualizado.getNombre_prod());
 			productoExistente.setPrecio_unitario(productoActualizado.getPrecio_unitario());
 			productoExistente.setCantidad(productoActualizado.getCantidad());
-			productoExistente.setCategoria(productoActualizado.getCategoria());
+			productoExistente.setIdCategoria(id);
 		}
 		
 		
@@ -84,16 +90,34 @@ public class ProductoControlador {
 	
 	//Eliminar
 	@DeleteMapping("producto/{id}")
-	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void eliminarProducto(@PathVariable Integer id) {
+	public ResponseEntity<?> eliminarProducto(@PathVariable Integer id) {
 		
+		Map<String, Object> response = new HashMap<>();
+		
+		try {
+				
+		if(!productoServicio.existePorId(id)) {
+			
+			response.put("mensaje", "El producto con ID " + id + " no existe.");	
+			return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);		
+		}
+		
+		//Elimina por Id
 		productoServicio.eliminarProducto(id);
 		
+		response.put("mensaje", "Producto eliminado");
+		return new ResponseEntity<>(response, HttpStatus.OK);
+		
+		
+		}catch(DataAccessException exDt) {
+			
+			response.put("mensaje", exDt.getMessage());
+			
+			response.put("producto", null);
+			
+			return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}	
 	}
-		
-
-	
-		
 }
